pyproject.toml
[project]
name = "my-app"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-allauth>=65.15.0",
    "django-browser-reload>=1.21.0",
    "django-cleanup>=9.0.0",
    "django-htmx>=1.27.0",
    "djangorestframework>=3.17.1",
    "pillow>=12.1.1",
]
